+++ packages/react-core/src/components/Select/index.ts
@@ -2,3 +2,4 @@ export * from './Select';
 export * from './SelectGroup';
 export * from './SelectList';
 export * from './SelectOption';
+export * from './wrapped/SelectTypeahead';
